refactor(credit): migrate external purchase statuses
